go: NewTransaction shouldn't specify the temp branch
Previously every call to NewTransaction needed to give a different
name for the temp branch since the branch namespace is shared. Clients
commonly used a constant string which would lead to concurrency failures
if the same function was called from 2 goroutines.

This patch removes the temporary branch name from the API and generates
a fresh one from an atomically-incremented counter. This still assumes
that only one client process can use this API at a time, but this is
better than assuming that only one Goroutine can use this at a time.

Signed-off-by: David Scott <dave.scott@docker.com>